Removing renderer folder complexities to use vike-vue features
