+++ src/pages/_layout/Head.vue
@@ -1,0 +1,6 @@
+<template>
+  <meta
+    name="description"
+    content="Test decription"
+  >
+</template>
