+++ src/pages/_layout/Layout.vue
@@ -1,20 +1,30 @@
 <template>
   <div id="page-container">
     <div id="sidebar">
-      <a href="/" id="logo-container">
-        <img src="./logo.svg" height="64" width="64" alt="logo" />
+      <a
+        id="logo-container"
+        href="/"
+      >
+        <img
+          src="./img/logo.svg"
+          height="64"
+          width="64"
+          alt="logo"
+        >
       </a>
       <Link href="/">Welcome</Link>
       <Link href="/about">About</Link>
       <Link href="/star-wars">Data Fetching</Link>
     </div>
-    <div id="page-content"><slot /></div>
+    <div id="page-content">
+      <slot />
+    </div>
   </div>
 </template>
 
 <script setup>
-import Link from './Link.vue'
-import './css/index.css'
+import Link from './Link.vue';
+import './css/index.css';
 </script>
 
 <style>
@@ -49,5 +59,18 @@ import './css/index.css'
 #logo-container {
   margin-top: 20px;
   margin-bottom: 10px;
+}
+</style>
+
+<style lang="scss">
+$back-color: #fcc;
+$font-color: #88c;
+
+body {
+  background-color: $back-color;
+
+  h1 {
+    color: $font-color;
+  }
 }
 </style>
