+++ src/pages/_layout/css/index.css
@@ -1,5 +1,5 @@
-@import 'bulma/css/bulma.css';
 @import './reset.css';
 @import './links.css';
 @import './code.css';
 @import './page-transition-loading-animation.css';
+@import 'bulma/css/bulma.css';
