+++ src/pages/index/+Page.vue
@@ -10,16 +10,3 @@
 <script setup>
 import Counter from './Counter.vue';
 </script>
-
-<style lang="scss">
-$back-color: #fcc;
-$font-color: #88c;
-
-body {
-  background-color: $back-color;
-
-  h1 {
-    color: $font-color;
-  }
-}
-</style>
