+++ src/pages/star-wars/@id/+data.js
@@ -4,6 +4,8 @@ import fetch from 'cross-fetch';
 export { data };
 
 const data = async (pageContext) => {
+    // simulate a slow network
+    await sleep(1000);
     const response = await fetch(`https://brillout.github.io/star-wars/api/films/${pageContext.routeParams.id}.json`);
     let movie = await response.json();
 
@@ -26,4 +28,8 @@ function minimize(movie) {
         id, title, release_date, director, producer,
     };
     return movie;
+}
+
+async function sleep(ms) {
+    return new Promise((resolve) => setTimeout(resolve, ms));
 }
