+++ src/pages/star-wars/index/+Page.vue
@@ -10,6 +10,6 @@
 </template>
 
 <script setup>
-import { useData } from '../../../renderer/useData'
+import { useData } from ':useData';
 const data = useData()
 </script>
