--- src/renderer/+config.h.js
@@ -1,17 +1,0 @@
-// https://vike.dev/config
-export default {
-    // https://vike.dev/clientRouting
-    clientRouting: true,
-    // https://vike.dev/meta
-    meta: {
-    // Define new setting 'title'
-        title: {
-            env: { server: true, client: true },
-        },
-        // Define new setting 'description'
-        description: {
-            env: { server: true },
-        },
-    },
-    hydrationCanBeAborted: true,
-};
