+++ src/renderer/+onPageTransitionEnd.js
@@ -1,7 +1,7 @@
-// https://vike.dev/onPageTransitionEnd
-export { onPageTransitionEnd };
-
 async function onPageTransitionEnd() {
     console.log('Page transition end');
     document.querySelector('body').classList.remove('page-is-transitioning');
 }
+
+/* eslint-disable import/prefer-default-export */
+export { onPageTransitionEnd };
