+++ src/renderer/+onPageTransitionStart.js
@@ -1,7 +1,7 @@
-// https://vike.dev/onPageTransitionStart
-export { onPageTransitionStart };
-
 function onPageTransitionStart() {
     console.log('Page transition start');
     document.querySelector('body').classList.add('page-is-transitioning');
 }
+
+/* eslint-disable import/prefer-default-export */
+export { onPageTransitionStart };
