--- src/renderer/useData.js
@@ -1,11 +1,0 @@
-// https://vike.dev/useData
-import { computed } from 'vue';
-import { usePageContext } from './usePageContext';
-
-export { useData };
-
-/** https://vike.dev/useData */
-function useData() {
-    const data = computed(() => usePageContext().data);
-    return data;
-}
