+++ vite.config.js
@@ -23,7 +23,9 @@ export default defineConfig({
     ],
     resolve: {
         alias: {
-            '#utils/logger': resolve(__dirname, 'src/js/utils/logger.js'),
+            ':utils/logger': resolve(__dirname, 'src/js/utils/logger.js'),
+            ':useData': 'vike-vue/useData',
+            ':usePageContext': 'vike-vue/usePageContext',
         },
     },
     server: {
